feat: support service account name for ingestion (#21)
* feat: support service account name for ingestion

This PR allows the service account name to be specified for crons. This
is useful if you have created specific accounts with permissions that
are needed to access the metadata (e.g. you might have a Kubernetes
Service Account that assumes an IAM Role with access to Glue for a Glue
Metadata Cronjob).

* build: bump chart version