feat(datahub-upgrade): Restore Indices Resources (#152)
* <feat>(datahub-upgrade): Restore Indices Resources

This PR adds the ability for end-users to specify resource requests and limits
for the datahub upgrade job 'restore indices'. This is needed for situations
where the restore indices job will OOM due to a large backfill of data. I've
seen this happening with our deployment of datahub and am unable to tune
the resources for the restore indices job because the current requests
and limits are hard-coded.