feat: add `monitoring.portName` in values.yaml
Add support for customizing the `portName` for monitoring.  It is currently hardcoded
to `jmx`.  In our setup, the prometheus scraper is looking for ports that are called
`prometheus`.  This change allows this value to be customized to any value that is
needed for a specific environment.  If no value is provided, it remains `jmx`.