feat: add custom image commands and arguments (#408)
* feat: add custom image commands and arguments

Add support for customizing an image's command or arguments.  In our setup, we have
some custom startup & shutdown behavior that need to run before the main process
starts for things like enabling / disabling networking or injecting secrets.  These
cannot be configured with init containers since they need to run in the same place
that the main process is going to run.  If command / args are empty or not present,
they will be excluded from helm, allowing the default command of the image to be used

---------

Co-authored-by: Travis Cook <tcook@slack-corp.com>